this is the intermediate assignment
This project is incomplete as I was burdened with family errands while working on this. However I have successfully created a game where a player is able to move, wall climb and jump as well as shoot fireballs. I have been unable due to time to add enemies and the pause menu. I therefore hope that this will suffice